--- src/main/java/com/GDGoC/BaS/user/Eye.java
@@ -1,4 +1,0 @@
-package com.GDGoC.BaS.user;
-
-public enum Eye {
-}
