--- src/main/java/com/GDGoC/BaS/user/Mouth.java
@@ -1,4 +1,0 @@
-package com.GDGoC.BaS.user;
-
-public enum Mouth {
-}
