+++ src/main/java/com/GDGoC/BaS/user/User.java
@@ -4,6 +4,10 @@ import com.GDGoC.BaS.booth.BoothUser;
 import com.GDGoC.BaS.drop.DropHistory;
 import com.GDGoC.BaS.notification.Notification;
 import com.GDGoC.BaS.shower.UserRecord;
+import com.GDGoC.BaS.user.enums.Eye;
+import com.GDGoC.BaS.user.enums.Mouth;
+import com.GDGoC.BaS.user.enums.Nose;
+import com.GDGoC.BaS.user.enums.Skin;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
