+++ src/main/java/com/GDGoC/BaS/user/enums/Nose.java
@@ -1,0 +1,4 @@
+package com.GDGoC.BaS.user.enums;
+
+public enum Nose {
+}
